Fix warnings int to size_t

+++ examples/Painter/MorphingJuliaFractal.cpp
@@ -23,11 +23,7 @@ int main()
     constexpr std::size_t height = 480;
     constexpr std::size_t bytesPerPixel = 3;
 
-    auto windowResult = SoftwareWindow::Create(
-        static_cast<int>(width),
-        static_cast<int>(height),
-        "Daily C++ Modules: Morphing Julia Fractal"
-    );
+    auto windowResult = SoftwareWindow::Create(width, height, "Daily C++ Modules: Morphing Julia Fractal");
 
     if (!windowResult)
     {
